+++ src/ponyc/main.c
@@ -189,6 +189,8 @@ static void usage()
     "                  point value. Defaults to 2.0.\n"
     "  --ponynoyield   Do not yield the CPU when no work is available.\n"
     "  --ponynoblock   Do not send block messages to the cycle detector.\n"
+    "  --ponypinasio   Pin the ASIO thread to a CPU the way Scheduler\n"
+    "                  threads are pinned to CPUs.\n"
     );
 }